requirements.txt
boto3==1.42.38
botocore==1.42.38
certifi==2026.1.4
charset-normalizer==3.4.4
dotenv==0.9.9
durationpy==0.10
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
kubernetes==35.0.0
Markdown==3.10.1
oauthlib==3.3.1
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
websocket-client==1.9.0
